requirements.txt
openai
langchain>=0.1.13
langchain-community
langchain-openai
langchain-huggingface
faiss-cpu
sentence-transformers>=2.2.2
python-docx
unstructured[all-docs]
pypdf
python-dotenv
nltk>=3.8.1
numpy
scikit-learn
rouge-score>=0.1.2
bert-score>=0.3.13
gradio
langchain-core
networkx
neo4j
langchain-neo4j
langfuse
requests
fastapi
uvicorn[standard]
pytest